#9 추가버튼 컴포넌트 분리
