+++ src/Todo.tsx
@@ -1,29 +1,13 @@
 import React, { useState } from 'react'
 
-import PortalModal from './PortalModal'
-import SampleModal from './SampleModal'
+import TodoAddBtn from './TodoAddBtn'
 import TodoList from './TodoList'
 
 const Todo = () => {
-  const [isModalOpen, isOpen] = useState(false)
-
-  const handleOpen = () => {
-    isOpen(true)
-  }
-
-  const handleClose = () => {
-    isOpen(false)
-  }
-
   return (
     <>
-      <button onClick={handleOpen}>Open Modal</button>
+      <TodoAddBtn />
       <TodoList />
-      {isModalOpen && (
-        <PortalModal>
-          <SampleModal closePop={handleClose} />
-        </PortalModal>
-      )}
     </>
   )
 }
