--- src/TodoAddModal.tsx
@@ -1,1 +1,0 @@
-import React from "react";
